Add files via upload
Created void method testSimpleGene

Created five test DNA strings with specific test cases, DNA with no “ATG”, DNA with no “TAA”, DNA with no “ATG” or “TAA”, DNA with ATG, TAA and the substring between them is a multiple of 3 (a gene), and DNA with ATG, TAA and the substring between them is not a multiple of 3. 

For each DNA string:
I called findSimpleGene with this sample DNA string as the parameter. If a gene exists following the algorithm above, then it prints the gene, otherwise it prints the empty string.